Initial implementation of marker/header removers
This is still a work in progress - it works but it needs a lot of manual
checking since there are often cases where the text of an ayah touches
the marker and that could cause the marker to be cut when not careful.
For checking, Kaleidoscope app is really useful since an overlay of the
photo with and without the ayah markers can help quickly pinpoint where
extra pixels were removed.

Will likely try to find a better way to write this in the future in sha'
Allah.